Checkpointing - entirely too many new features/changes to enumerate, but in general:   - new concise, tab-delimited output format with edit strings   - lots of little fixes due to valgrind and GCC/ICC -Wall   - efficiency fixes to prettyprint   - calculation fixes for probcalc (Marc)   - probcalc support for independent (parallel) rates + probability calcs   - prettyprint/probcalc/mergehits conversion to new fasta API   - phil bugs squashed (gpre/gpost bogus in RC, bogus fringe sequences)   - sort and nuke duplicates in rmapper   - zlib support and a _fast_ zlib-backed fgets replacement (loading files     is much faster now)   - documentation updates (incomplete)   - makefile hackery to get flags right and avoid -static on OS X
git-svn-id: svn+ssh://combine-101/home/svn_bio/trunk@245 450597fd-3c47-0410-84d8-f85f2794411d

+++ common/dag_glue.cpp
@@ -6,6 +6,7 @@
 #include <string>
 #include <climits>
 #include <fstream>
+#include <zlib.h>
 
 #include "../common/dag_align.h"
 #include "../common/dag_glue.h"
@@ -89,7 +90,7 @@ dag_get_kmers(dag_cookie_t dct, int len)
 
 	ki = kg->getkmers(len);
 
-	if ((len - 1) >= (ki->kmers).size()) {
+	if ((len - 1) >= (int)(ki->kmers).size()) {
 		delete ki;
 		return (NULL);
 	}
